Initial migration to `VectorSimilarityTable`
