--- src/components/TickerDetail/TickerDetail.TickerVectorTable/index.tsx
@@ -1,6 +1,0 @@
-import TickerVectorTable, {
-  TickerVectorTableProps,
-} from "./TickerDetail.TickerVectorTable";
-
-export default TickerVectorTable;
-export type { TickerVectorTableProps };
