+++ src/components/VectorSimilarityTable/index.tsx
@@ -1,0 +1,6 @@
+import VectorSimilarityTable, {
+  VectorSimiliartyTableProps,
+} from "./VectorSimilarityTable";
+
+export default VectorSimilarityTable;
+export type { VectorSimiliartyTableProps };
